chore: make state transitions uncancelable
Ensures updates to `KafkaConsumerActor`'s internal state are applied
together with the side-effects that support those updates by executing
both together inside an uncancelable block. These changes ensure that
changes made to the consumer are reflected in the internal state, and
changes to the internal state are made visible to clients.

In the simplest cases, a sequence of `Ref.update`/`modify` and
`Async.flatMap` was refactored to use `Ref.flatModify`, which offers the
desired guarantees. In other cases, the relevant actions were explicitly
wrapped in an `Async.uncancelable` block.

Where relevant, logging of state changes was treated as part of the
uncancelable block. This is to avoid silent state changes, which may be
hard to track down while debugging. Invocation of user-supplied
callbacks on changes to assigned partitions is left out of these blocks
as failures in those do not affect the internal state of the consumer.