chore: create underlying KafkaConsumer in the Blocking context
The use of WithConsumer ensures that consumer methods are called from
the Blocking context, but it left out the initial instantiation of the
consumer instance.

With this change, the constructor is also called in the same Blocking
context.

This addresses an issue originally identified by @eugkhp in
https://github.com/fd4s/fs2-kafka/pull/1364, where the constructor is
observed to make network calls that should not happen in a non-blocking
context.